Pass domain and port separately and fix letsenc.
This allows to add the correct port to the security group.
Add some logic to autogenerate the public_url fro it, but
allow it to be overriden.

letsenc_* were not passed into the script previously, this is fixed now.

If letsenc_mail is set, we will try to generate Let's Encrypt
certificates on the fly now. cert.* still needs to exist, but won't get
used (and can thus be empty files). Let's Encrypt is untested thus far.