Rename UIlanguage to ui_language, hide tweak_aspectRatio.
aspectRatio is no longer included in latest templates, so let's not play
with it. (Code is still there, but comment in code and no longer
documented.)
UIlanguage was not matching our option naming, adjust to ui_language.
We also changed a min: parameter after framerate for screen sharing
which was commented out, avoid this now.