Avoid waiting indefinitely for pipstrap.py.
When building the web container in docker, bootstrapping the certbot
does invoke pipstrap to download packages from pypi.
For reasons not yet completely debugged, this hangs in betacloud --
the connection gets established but then never responds to the first
data exchange (TLS handshake).

This is harmless if we inject certs and are not using the certbot to
actually get new certs from letsenc. So in this case, let's just continue
after 42s if the pistrap did not succeed.