Allow using existing network.
We may not have sufficient quota or permissions to create a new router
and a new network+subnet to deploy our jitsi instance. So we allow the
user to specify an existing network in net_jitsi:

This is currently on a branch -- I have yet to find an elegant way to
enable both modes of operation.