Make Jitsi come up again after a reboot.
To achieve this, split jitsi bootstrap into 3 parts.

The last part is the only piece required to restart after a reboot.
(Jitsi is strange in that it replaces the web/config.js after each
container restart, so we need to tweak settings and restart nginx.)

Add boot.local script to include the start/stop for jitsi.

Signed-off-by: Kurt Garloff <kurt@garloff.de>