Add JIGASI_PORT_MIN/MAX if missing. Unnice JVB path.
Previously, there were defaults in the ~/.env file for JIGASI_PORT_MIN/
MAX that we just tweaked. Nowadays, they are not there. So add as needed.
This saves the scheduling/memory footprint of ~30 docker-proxy processes
(with our default of 20020 as JIGASI_PORT_MAX).

Also run jvb and the docker-proxy on port 10000 with higher priority
(niceness of -5) -- not because we expect these to be CPU hogs, but
because we want them to preempt other tasks to lower latency.

Signed-off-by: Kurt Garloff <kurt@garloff.de>